% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HAI_data.R
\docType{data}
\name{HAI_data}
\alias{HAI_data}
\title{German Healthcare-Associated Infections Data}
\format{
A data frame with 6 rows and 6 variables:
\describe{
\item{Infection}{Type of HAI (HAP, SSI, BSI, UTI, CDI, ALL)}
\item{Cases}{Annual number of cases}
\item{Deaths}{Annual number of attributable deaths}
\item{DALY}{Annual DALYs}
\item{YLL}{Years of life lost}
\item{YLD}{Years lived with disability}
}
}
\source{
German Point Prevalence Survey 2011
}
\usage{
HAI_data
}
\description{
This dataset contains annual cases, deaths, DALYs, YLLs and YLDs for five types of HAIs in Germany (2011).
}
\keyword{datasets}
